+++ src/probot/devices/motors/motor.hpp
@@ -5,7 +5,7 @@ namespace probot::motor {
   struct IMotor {
     virtual bool claim(void* owner) = 0;                    // exclusive write claim
     virtual void release(void* owner) = 0;                  // release claim if owner matches
-    virtual bool setPower(int16_t power, void* owner) = 0;  // -1000..1000; only current owner succeeds
+    virtual bool setPower(float power, void* owner) = 0;     // -1.0..1.0 normalized output
     virtual bool isClaimed() const = 0;
     virtual void* currentOwner() const = 0;
